move fuzz tests. Got rid of global test_data_path
