+++ CMakeLists.txt
@@ -215,6 +215,7 @@ set(FlatBuffers_Tests_SRCS
   src/idl_gen_fbs.cpp
   tests/evolution_test.cpp
   tests/flexbuffers_test.cpp
+  tests/fuzz_test.cpp
   tests/json_test.cpp
   tests/monster_test.cpp
   tests/optional_scalars_test.cpp
